Move Fiji UI layer from ome.zarr.fiji.ui to ome.zarr.fijiui
Promote the UI layer to a top-level sibling of ome.zarr.fiji (rather than
a sub-package of it), making the future ome-zarr-fiji-ui artifact boundary
more explicit.
