--- src/main/java/ome/zarr/fiji/ui/plugin/OpenInBDVCommand.java
@@ -1,72 +1,0 @@
-/*-
- * #%L
- * OME-Zarr extras for Fiji
- * %%
- * Copyright (C) 2022 - 2026 SciJava developers
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-package ome.zarr.fiji.ui.plugin;
-
-import java.lang.invoke.MethodHandles;
-
-import org.scijava.command.Command;
-import org.scijava.plugin.Parameter;
-import org.scijava.plugin.Plugin;
-import org.scijava.ui.UIService;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import ome.zarr.fiji.Pyramidal;
-import ome.zarr.fiji.PyramidalBdv;
-import ome.zarr.fiji.plugins.PyramidalService;
-import ome.zarr.fiji.util.BdvUtils;
-
-@Plugin( type = Command.class, menuPath = "Plugins > OME-Zarr > Open Current OME-Zarr Image in BigDataViewer" )
-public class OpenInBDVCommand implements Command
-{
-	private static final Logger logger = LoggerFactory.getLogger( MethodHandles.lookup().lookupClass() );
-
-	@Parameter
-	private UIService uiService;
-
-	@Parameter( required = false )
-	private PyramidalService pyramidalService;
-
-	@Parameter( required = false )
-	private Pyramidal pyramidal;
-
-	@Override
-	public void run()
-	{
-		logger.trace( "Running OpenInBDVCommand. pyramidal={}", pyramidal );
-		if ( pyramidal == null )
-		{
-			uiService.showDialog( "The active image is not an OME-Zarr dataset.", "Open in BigDataViewer" );
-			return;
-		}
-		final PyramidalBdv< ? > bdvDataset = new PyramidalBdv<>( pyramidal.getContext(), pyramidal.getPyramidContents() );
-		BdvUtils.showBdvAndRegisterDataset( bdvDataset, pyramidalService );
-	}
-}
