+++ src/main/java/ome/zarr/fijiui/dialog/CreateIcon.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,54 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.dialog;
 
-import net.imagej.ImageJ;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import javax.swing.ImageIcon;
+import java.awt.Image;
+import java.lang.invoke.MethodHandles;
+import java.net.URL;
 
-import java.net.URI;
-import java.nio.file.Paths;
-
-public class ScriptUtilsDemo
+public class CreateIcon
 {
-	public static void main( String[] args )
-	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	private static final Logger logger = LoggerFactory.getLogger( MethodHandles.lookup().lookupClass() );
+
+	private CreateIcon()
+	{
+		// prevent instantiation
+	}
+
+	/**
+	 * Loads an image from the classpath (e.g., in src/main/resources)
+	 * and scales it to 32x32 pixels.
+	 *
+	 * @param resourcePath path to the image relative to the classpath,
+	 *                     e.g. "/icons/myicon.png"
+	 * @return a scaled ImageIcon, or an empty one if loading fails
+	 */
+	public static ImageIcon getAndResizeIcon( final String resourcePath )
+	{
+
+		try
+		{
+			URL resourceUrl = CreateIcon.class.getResource( resourcePath );
+			if ( resourceUrl == null )
+			{
+				throw new IllegalArgumentException( "Resource not found: " + resourcePath );
+			}
+
+			Image image = new ImageIcon( resourceUrl ).getImage().getScaledInstance( 32, 32, Image.SCALE_SMOOTH );
+
+			return new ImageIcon( image );
+		}
+		catch ( Exception e )
+		{
+			logger.debug( "Failed to load icon from path: " + resourcePath, e );
+			// Fallback to an empty image if the resource can't be loaded
+			return new ImageIcon();
+		}
 	}
 }
