+++ src/main/java/ome/zarr/fijiui/open/options/ZarrOpenBehavior.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,60 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.open.options;
 
-import net.imagej.ImageJ;
+import java.util.NoSuchElementException;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
-
-import java.net.URI;
-import java.nio.file.Paths;
-
-public class ScriptUtilsDemo
+/**
+ * Options for opening OME-Zarr datasets in different viewers and resolutions.
+ */
+public enum ZarrOpenBehavior
 {
-	public static void main( String[] args )
-	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
+	/**
+	 * Open the highest available single-resolution in ImageJ.
+	 */
+	IMAGEJ_HIGHEST_RESOLUTION( "Open the highest available single-resolution in ImageJ" ),
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	/**
+	 * Open a matching single-resolution image in ImageJ.
+	 */
+	IMAGEJ_CUSTOM_RESOLUTION( "Open a matching single-resolution image in ImageJ" ),
+
+	/**
+	 * Open as multi-resolution in BigDataViewer (BDV).
+	 */
+	BDV_MULTI_RESOLUTION( "Open as a multi-resolution source in BigDataViewer" ),
+
+	/**
+	 * Always open the selection dialog with different icons.
+	 */
+	SHOW_SELECTION_DIALOG( "Open selection dialog with choices" );
+
+	private final String description;
+
+	ZarrOpenBehavior( final String description )
+	{
+		this.description = description;
+	}
+
+	public static ZarrOpenBehavior getByName( final String name )
+	{
+		for ( final ZarrOpenBehavior option : values() )
+			if ( option.name().equals( name ) )
+				return option;
+		throw new NoSuchElementException( name );
+	}
+
+	public static ZarrOpenBehavior getByDescription( final String description )
+	{
+		for ( final ZarrOpenBehavior behavior : values() )
+			if ( behavior.description.equals( description ) )
+				return behavior;
+		return null;
+	}
+
+	public String getDescription()
+	{
+		return description;
 	}
 }
