+++ src/main/java/ome/zarr/fijiui/plugin/PasteOmeZarrUrlCommand.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,30 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.plugin;
 
-import net.imagej.ImageJ;
+import org.scijava.Context;
+import org.scijava.command.Command;
+import org.scijava.plugin.Parameter;
+import org.scijava.plugin.Plugin;
 
 import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import ome.zarr.fijiui.open.PasteToOpenAction;
 
-import java.net.URI;
-import java.nio.file.Paths;
-
-public class ScriptUtilsDemo
+/**
+ * Reads a URI from the system clipboard and opens it as an OME-Zarr
+ * dataset, using the same backend, resolution, and open-behavior settings as
+ * the drag-and-drop pipeline. Mirrors napari's "paste URI to open" UX.
+ */
+@Plugin( type = Command.class, menuPath = "Plugins > OME-Zarr > Paste OME-Zarr URI" )
+public class PasteOmeZarrUrlCommand implements Command
 {
-	public static void main( String[] args )
-	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
+	@Parameter
+	private Context context;
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	@Override
+	public void run()
+	{
+		PasteToOpenAction.pasteFromClipboard( context, IJ::error );
 	}
 }
