+++ src/main/java/ome/zarr/fijiui/settings/OpeningBehaviorSettings.java
@@ -1,0 +1,147 @@
+/*-
+ * #%L
+ * OME-Zarr extras for Fiji
+ * %%
+ * Copyright (C) 2022 - 2026 SciJava developers
+ * %%
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ * #L%
+ */
+package ome.zarr.fijiui.settings;
+
+import java.lang.invoke.MethodHandles;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.scijava.ItemVisibility;
+import org.scijava.command.Command;
+import org.scijava.command.DynamicCommand;
+import org.scijava.plugin.Parameter;
+import org.scijava.plugin.Plugin;
+import org.scijava.prefs.PrefService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import ome.zarr.fijiui.open.options.ZarrOpeningSettings;
+import ome.zarr.fijiui.open.options.ZarrOpenBehavior;
+import ome.zarr.fiji.open.ZarrReaderBackend;
+
+/**
+ * A FIJI/ImageJ command to select what to do when an OME-Zarr image is Drag &amp; Dropped into Fiji.
+ */
+@Plugin( type = Command.class, menuPath = "Plugins > OME-Zarr > Settings > Opening Behavior Settings", initializer = "init" )
+public class OpeningBehaviorSettings extends DynamicCommand
+{
+	private static final Logger logger = LoggerFactory.getLogger( MethodHandles.lookup().lookupClass() );
+
+	private static final Integer WIDTH = 20;
+
+	@SuppressWarnings( "all" )
+	@Parameter
+	private PrefService prefService;
+
+	@SuppressWarnings( "all" )
+	@Parameter( visibility = ItemVisibility.MESSAGE, required = false, persist = false )
+	private String infoMessage = "<html>"
+			+ "<body width=" + WIDTH + "cm align=left>"
+			+ "Configure the behavior when OME-Zarr datasets are drag & dropped (local folders) or copy & pasted (local and remote paths) into Fiji.<br>"
+			+ "Choose a default behavior, optionally limit the preferred resolution and choose the reader backend libary."
+			+ "</body>"
+			+ "</html>";
+
+	@SuppressWarnings( "all" )
+	@Parameter( label = "Default opening behavior", description = "Choose the opening behavior if you drag & drop local OME-Zarr folders or copy & paste OME-Zarr paths (local/remote) into Fiji", initializer = "initZarrOpenBehaviors" )
+	private String defaultZarrOpenBehavior;
+
+	@SuppressWarnings( "all" )
+	@Parameter( label = "Preferred width for 'matching resolution' choice" )
+	private int preferredWidth;
+
+	@SuppressWarnings( "all" )
+	@Parameter( visibility = ItemVisibility.MESSAGE, required = false, persist = false )
+	private String preferredWidthInfo = "<html>"
+			+ "<body width=" + WIDTH + "cm align=left>"
+			+ "For the 'matching resolution' behavior, set a preferred maximum width.<br>"
+			+ "Fiji will open the highest available resolution whose width is below this value.<br>"
+			+ "If no such resolution exists, the image will not be opened."
+			+ "</body>"
+			+ "</html>";
+
+	@SuppressWarnings( "all" )
+	@Parameter( label = "Reader backend", description = "Choose which library is used to read OME-Zarr datasets", initializer = "initZarrReaderBackends" )
+	private String readerBackendChoice;
+
+	@SuppressWarnings( "all" )
+	@Parameter( visibility = ItemVisibility.MESSAGE, required = false, persist = false )
+	private String readerBackendInfo = "<html>"
+			+ "<body width=" + WIDTH + "cm align=left>"
+			+ "N5 supports OME-Zarr v0.4 (Zarr v2) - v0.5 (Zarr v3).<br>"
+			+ "zarr-java supports OME-Zarr v0.4 (Zarr v2) - v0.3 (Zarr v3)."
+			+ "</body>"
+			+ "</html>";
+
+	private ZarrOpeningSettings settings;
+
+	@Override
+	public void run()
+	{
+		settings.setCurrentChoice( ZarrOpenBehavior.getByDescription( defaultZarrOpenBehavior ) );
+		settings.setPreferredMaxWidth( preferredWidth );
+		settings.setReaderBackend( ZarrReaderBackend.getByDescription( readerBackendChoice ) );
+		logger.debug( "Now saving OME-Zarr settings to user preferences. Behavior: {}, preferredWidth: {}, readerBackend: {}",
+				settings.getOpenBehavior(), preferredWidth, settings.getReaderBackend() );
+		settings.saveSettingsToPreferences( prefService );
+	}
+
+	@SuppressWarnings( "unused" )
+	private void init()
+	{
+		settings = ZarrOpeningSettings.loadSettingsFromPreferences( prefService );
+		defaultZarrOpenBehavior = settings.getOpenBehavior().getDescription();
+		preferredWidth = settings.getPreferredMaxWidth();
+		readerBackendChoice = settings.getReaderBackend().getDescription();
+	}
+
+	@SuppressWarnings( "unused" )
+	private void initZarrOpenBehaviors()
+	{
+		getInfo().getMutableInput( "defaultZarrOpenBehavior", String.class ).setChoices( enumNamesAsList( ZarrOpenBehavior.values() ) );
+	}
+
+	@SuppressWarnings( "unused" )
+	private void initZarrReaderBackends()
+	{
+		getInfo().getMutableInput( "readerBackendChoice", String.class )
+				.setChoices( readerBackendDescriptions( ZarrReaderBackend.values() ) );
+	}
+
+	static List< String > enumNamesAsList( final ZarrOpenBehavior[] values )
+	{
+		return Arrays.stream( values ).map( ZarrOpenBehavior::getDescription ).collect( Collectors.toList() );
+	}
+
+	static List< String > readerBackendDescriptions( final ZarrReaderBackend[] values )
+	{
+		return Arrays.stream( values ).map( ZarrReaderBackend::getDescription ).collect( Collectors.toList() );
+	}
+}
