+++ src/main/java/ome/zarr/fijiui/settings/UserScriptSettings.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,32 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.settings;
 
-import net.imagej.ImageJ;
+import org.scijava.command.Command;
+import org.scijava.log.LogService;
+import org.scijava.plugin.Parameter;
+import org.scijava.plugin.Plugin;
+import org.scijava.widget.FileWidget;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import java.io.File;
 
-import java.net.URI;
-import java.nio.file.Paths;
-
-public class ScriptUtilsDemo
+@Plugin( type = Command.class, menuPath = "Plugins > OME-Zarr > Settings > User Script Settings", name = "OMEZarrUserScript", headless = true )
+public class UserScriptSettings implements Command
 {
-	public static void main( String[] args )
-	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	@Parameter( label = "Tooltip text:" )
+	String scriptTitle;
+
+	@Parameter( label = "Script to be executed:", style = FileWidget.OPEN_STYLE )
+	File scriptPath;
+
+	@Parameter
+	LogService log;
+
+	@Override
+	public void run()
+	{
+		log.info( "Thanks, memorizing the path: " + scriptPath );
 	}
 }
