--- src/test/java/ome/zarr/fiji/ui/dialog/DnDActionChooserTest.java
@@ -1,72 +1,0 @@
-/*-
- * #%L
- * OME-Zarr extras for Fiji
- * %%
- * Copyright (C) 2022 - 2026 SciJava developers
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-package ome.zarr.fiji.ui.dialog;
-
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assumptions.assumeFalse;
-
-import java.awt.GraphicsEnvironment;
-
-import javax.swing.JDialog;
-import javax.swing.SwingUtilities;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-class DnDActionChooserTest
-{
-	@BeforeEach
-	void requireDisplay()
-	{
-		assumeFalse( GraphicsEnvironment.isHeadless(), "Skipped in headless environment" );
-	}
-
-	@Test
-	void showDialogDoesNotThrow()
-	{
-		assertDoesNotThrow( () -> new DnDActionChooser( null, null ) );
-	}
-
-	@Test
-	void showDialogCreatesVisibleDialog() throws Exception
-	{
-		DnDActionChooser chooser = new DnDActionChooser( null, null );
-		SwingUtilities.invokeAndWait( chooser::showDialog );
-
-		JDialog dialog = chooser.currentDialog;
-		assertNotNull( dialog, "showDialog should have created a JDialog" );
-		assertTrue( dialog.isShowing() );
-
-		SwingUtilities.invokeAndWait( chooser::dispose );
-		assertFalse( dialog.isShowing() );
-	}
-}
