--- src/test/java/ome/zarr/fiji/ui/open/PasteToOpenActionTest.java
@@ -1,68 +1,0 @@
-/*-
- * #%L
- * OME-Zarr extras for Fiji
- * %%
- * Copyright (C) 2022 - 2026 SciJava developers
- * %%
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * 
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- * #L%
- */
-package ome.zarr.fiji.ui.open;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.awt.Toolkit;
-import java.awt.datatransfer.StringSelection;
-import java.net.URISyntaxException;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Consumer;
-
-import org.junit.jupiter.api.Test;
-import org.scijava.Context;
-
-import ome.zarr.ZarrTestUtils;
-
-class PasteToOpenActionTest
-{
-	private final List< String > errors = new ArrayList<>();
-
-	private final Consumer< String > errorHandler = errors::add;
-
-	@Test
-	void nonZarrLocalPathReportsError() throws URISyntaxException
-	{
-		try (Context context = new Context())
-		{
-			final Path path = ZarrTestUtils.resourcePath( "ome/zarr/testdata/2d_testing" );
-			Toolkit.getDefaultToolkit().getSystemClipboard()
-					.setContents( new StringSelection( path.toString() ), null );
-			boolean result = PasteToOpenAction.pasteFromClipboard( context, errorHandler );
-			assertFalse( result );
-			assertEquals( 1, errors.size() );
-			assertTrue( errors.get( 0 ).contains( "OME-Zarr" ) );
-		}
-	}
-}
