+++ src/test/java/ome/zarr/fijiui/dialog/DnDActionChooserTest.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,47 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.dialog;
 
-import net.imagej.ImageJ;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import java.awt.GraphicsEnvironment;
 
-import java.net.URI;
-import java.nio.file.Paths;
+import javax.swing.JDialog;
+import javax.swing.SwingUtilities;
 
-public class ScriptUtilsDemo
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class DnDActionChooserTest
 {
-	public static void main( String[] args )
+	@BeforeEach
+	void requireDisplay()
 	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
+		assumeFalse( GraphicsEnvironment.isHeadless(), "Skipped in headless environment" );
+	}
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	@Test
+	void showDialogDoesNotThrow()
+	{
+		assertDoesNotThrow( () -> new DnDActionChooser( null, null ) );
+	}
+
+	@Test
+	void showDialogCreatesVisibleDialog() throws Exception
+	{
+		DnDActionChooser chooser = new DnDActionChooser( null, null );
+		SwingUtilities.invokeAndWait( chooser::showDialog );
+
+		JDialog dialog = chooser.currentDialog;
+		assertNotNull( dialog, "showDialog should have created a JDialog" );
+		assertTrue( dialog.isShowing() );
+
+		SwingUtilities.invokeAndWait( chooser::dispose );
+		assertFalse( dialog.isShowing() );
 	}
 }
