+++ src/test/java/ome/zarr/fijiui/open/PasteToOpenActionTest.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,43 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.open;
 
-import net.imagej.ImageJ;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import java.awt.Toolkit;
+import java.awt.datatransfer.StringSelection;
+import java.net.URISyntaxException;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Consumer;
 
-import java.net.URI;
-import java.nio.file.Paths;
+import org.junit.jupiter.api.Test;
+import org.scijava.Context;
 
-public class ScriptUtilsDemo
+import ome.zarr.ZarrTestUtils;
+
+class PasteToOpenActionTest
 {
-	public static void main( String[] args )
-	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
+	private final List< String > errors = new ArrayList<>();
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+	private final Consumer< String > errorHandler = errors::add;
+
+	@Test
+	void nonZarrLocalPathReportsError() throws URISyntaxException
+	{
+		try (Context context = new Context())
+		{
+			final Path path = ZarrTestUtils.resourcePath( "ome/zarr/testdata/2d_testing" );
+			Toolkit.getDefaultToolkit().getSystemClipboard()
+					.setContents( new StringSelection( path.toString() ), null );
+			boolean result = PasteToOpenAction.pasteFromClipboard( context, errorHandler );
+			assertFalse( result );
+			assertEquals( 1, errors.size() );
+			assertTrue( errors.get( 0 ).contains( "OME-Zarr" ) );
+		}
 	}
 }
