+++ src/test/java/ome/zarr/fijiui/plugin/DnDHandlerPluginTest.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,27 +26,41 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package ome.zarr.examples.demo;
+package ome.zarr.fijiui.plugin;
 
-import net.imagej.ImageJ;
+import static org.mockito.Mockito.times;
 
-import ij.IJ;
-import ome.zarr.fijiui.util.ScriptUtils;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Path;
 
-import java.net.URI;
-import java.nio.file.Paths;
+import org.junit.jupiter.api.Test;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
+import org.scijava.Context;
+import org.scijava.io.location.FileLocation;
 
-public class ScriptUtilsDemo
+import ome.zarr.fijiui.open.ZarrOpenActions;
+import ome.zarr.ZarrTestUtils;
+
+class DnDHandlerPluginTest
 {
-	public static void main( String[] args )
+	@Test
+	void openDelegatesToZarrOpenActions() throws URISyntaxException, IOException
 	{
-		ImageJ ij = new ImageJ();
-		ij.ui().showUI();
+		try (Context context = new Context())
+		{
+			final Path path = ZarrTestUtils.resourcePath( "ome/zarr/testdata/2d_testing/2d_dataset_v4.ome.zarr/" );
+			final FileLocation fileLocation = new FileLocation( path.toUri() );
 
-		final URI inputUri = Paths
-				.get( "/home/ulman/Documents/talks/CEITEC/2025_11_ZarrSymposium_Zurich/data/MitoEM_fixedRes.zarr/MitoEM_fixedRes" ).toUri();
-		System.out.println( "\nLet's run the script... on URI param: " + inputUri );
-		ScriptUtils.executePresetScript( ij.context(), inputUri, IJ::error );
-		System.out.println( "Done." );
+			final DnDHandlerPlugin dnDHandlerPlugin = new DnDHandlerPlugin();
+			dnDHandlerPlugin.setContext( context );
+
+			try (MockedStatic< ZarrOpenActions > mocked = Mockito.mockStatic( ZarrOpenActions.class ))
+			{
+				dnDHandlerPlugin.open( fileLocation );
+				mocked.verify( () -> ZarrOpenActions.openWithSettings( path.toUri(), context ), times( 1 ) );
+			}
+		}
 	}
 }
